refactor: new file for errors
