+++ app.js
@@ -17,9 +17,10 @@ const { errors } = require('celebrate');
 // const { login, createUser } = require('./controllers/users');
 // const auth = require('./middlewares/auth');
 // const NotFoundError = require('./errors/NotFoundError');
-const { serverErrorMessage, serverFallMessage } = require('./configs/messages');
+const { serverFallMessage } = require('./configs/messages');
 const { limiter } = require('./configs/rateLimit');
 const { requestLogger, errorLogger } = require('./middlewares/logger');
+const { errorHandler } = require('./middlewares/errorHandler');
 
 // const { PORT = 3000 } = process.env;
 const { PORT, DB_URL } = require('./configs/index');
@@ -74,13 +75,14 @@ app.use(router);
 app.use(errorLogger);
 app.use(errors());
 
-app.use((err, req, res, next) => {
-  const { statusCode = 500, message } = err;
-  res.status(statusCode).send({
-    // message: statusCode === 500 ? 'На сервере произошла ошибка' : message,
-    message: statusCode === 500 ? serverErrorMessage : message,
-  });
-});
+// app.use((err, req, res, next) => {
+//   const { statusCode = 500, message } = err;
+//   res.status(statusCode).send({
+//     // message: statusCode === 500 ? 'На сервере произошла ошибка' : message,
+//     message: statusCode === 500 ? serverErrorMessage : message,
+//   });
+// });
+app.use(errorHandler);
 
 mongoose.connect(DB_URL, {
   useNewUrlParser: true,
