+++ routes/index.js
@@ -1,7 +1,11 @@
 /* eslint-disable no-unused-vars */
 const router = require('express').Router();
 
-const { celebrate, Joi } = require('celebrate');
+// const { celebrate, Joi } = require('celebrate');
+const {
+  validateUserLogin,
+  validateUserSignup,
+} = require('../middlewares/celebrateHandlers');
 const NotFoundError = require('../errors/NotFoundError');
 const { notFoundMessage } = require('../configs/messages');
 
@@ -11,20 +15,20 @@ const { login, createUser } = require('../controllers/users');
 
 const auth = require('../middlewares/auth');
 
-const validateUserLogin = celebrate({
-  body: Joi.object().keys({
-    email: Joi.string().required().email(),
-    password: Joi.string().required(),
-  }),
-});
+// const validateUserLogin = celebrate({
+//   body: Joi.object().keys({
+//     email: Joi.string().required().email(),
+//     password: Joi.string().required(),
+//   }),
+// });
 
-const validateUserSignup = celebrate({
-  body: Joi.object().keys({
-    email: Joi.string().required().email(),
-    password: Joi.string().required(),
-    name: Joi.string().required().min(2).max(30),
-  }),
-});
+// const validateUserSignup = celebrate({
+//   body: Joi.object().keys({
+//     email: Joi.string().required().email(),
+//     password: Joi.string().required(),
+//     name: Joi.string().required().min(2).max(30),
+//   }),
+// });
 
 router.post('/signin', validateUserLogin, login);
 router.post('/signup', validateUserSignup, createUser);
